Move the meta path finders in constants since it depends on winutil in particular
